chore: 코드 가독성을 위해 service단 코드 분리 + 중복되는 코드 함수화
