+++ backend/server/src/main/java/com/newengineeringghost/domain/api/controller/ApiController.java
@@ -2,7 +2,7 @@ package com.newengineeringghost.domain.api.controller;
 
 import com.newengineeringghost.domain.api.dto.ResponseDataDto;
 import com.newengineeringghost.domain.api.service.ApiService;
-import org.apache.tomcat.util.json.JSONParser;
+import com.newengineeringghost.domain.api.service.MongoDbService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,8 +16,13 @@ import java.util.Map;
 public class ApiController {
     private ApiService apiService;
 
+    private MongoDbService mongoDbService;
+
     @Autowired
     public void setApiService(ApiService apiService) {this.apiService = apiService;}
+
+    @Autowired
+    public void setMongoDbService(MongoDbService mongoDbService) {this.mongoDbService = mongoDbService;}
 
     // 빠른 측정
     @PostMapping("/quick")
@@ -52,12 +57,12 @@ public class ApiController {
     // mongodb 통신 test
     @GetMapping("/get")
     public ResponseDataDto getMeasurement(@RequestParam String link) throws IOException {
-        return apiService.getData(link);
+        return mongoDbService.getData(link);
     }
 
     // mongodb 통신 test
     @PostMapping("/post")
     public String postMeasurement(@RequestBody ResponseDataDto responseDto) {
-        return apiService.postData(responseDto);
+        return mongoDbService.postData(responseDto);
     }
 }
